consolidated interface
